\name{mlt-package}
\docType{package}
\alias{mlt-package}
\title{General Information on the \pkg{mlt} Package}
\description{
  The \pkg{mlt} package implements maximum likelihood estimation in 
  conditional transformation models as introduced by \bibcitet{mlt::Hothorn:Moest:Buehlmann:2017},
  its multivariate extension by \bibcitet{mlt::Klein:Hothorn:Barbanti:2020}, and
  location-scale models by \bibcitet{Siegfried_Kook_Hothorn_2023}.

  An introduction to the package is available in the \code{mlt} package
  vignette from package \code{mlt.docreg} \bibcitep{mlt::Hothorn:2018}.

  Novice users might find the high(er) level interfaces offered by package
  \pkg{tram} more convenient.

}
\references{\bibshow{*}}
\keyword{package}
