\name{mlt-package}
\docType{package}
\alias{mlt-package}
\title{General Information on the \pkg{mlt} Package}
\description{
  The \pkg{mlt} package implements maximum likelihood estimation in 
  conditional transformation models as introduced by Hothorn et al. (2020).

  An introduction to the package is available in the \code{mlt} package
  vignette from package \code{mlt.docreg} (Hothorn, 2020).

  A short talk on most likely transformations is available from
  \url{https://channel9.msdn.com/Events/useR-international-R-User-conference/useR2016/Most-Likely-Transformations}.

  Novice users might find the high(er) level interfaces offered by package
  \pkg{tram} more convenient.

}
\references{

  Torsten Hothorn, Lisa Moest, Peter Buehlmann (2018), Most Likely
  Transformations, \emph{Scandinavian Journal of Statistics}, \bold{45}(1),
  110--134, \doi{10.1111/sjos.12291}.

  Torsten Hothorn (2020), Most Likely Transformations: The mlt Package,
  \emph{Journal of Statistical Software}, \bold{92}(1), 1--68,
  \doi{10.18637/jss.v092.i01}

}
\author{
  This package is authored by Torsten Hothorn <Torsten.Hothorn@R-project.org>.
}
\keyword{package}
