\name{mlt-package}
\docType{package}
\alias{mlt-package}
\title{General Information on the \pkg{mlt} Package}
\description{
  The \pkg{mlt} implements maximum likelihood estimation in 
  conditional transformation models as introduced by Hothorn et al. (2017).

  An introduction to the package is available in the \code{mlt} package
  vignette from package \code{mlt.docreg}.
}
\references{

  Torsten Hothorn, Lisa Moest, Peter Buehlmann (2017), Most Likely
  Transformations, \emph{Scandinavian Journal of Statistics}, Accepted
  2017-06-19, \url{http://arxiv.org/abs/1508.06749}.

}
\author{
  This package is authored by Torsten Hothorn <Torsten.Hothorn@R-project.org>.
}
\keyword{package}
