\name{mlt-package}
\docType{package}
\alias{mlt-package}
\title{General Information on the \pkg{mlt} Package}
\description{
  The \pkg{mlt} implements maximum likelihood estimation in 
  conditional transformation models as introduced by Hothorn et al. (2016).

  An introduction to the package is available in the \code{mlt} package
  vignette from package \code{mlt.docreg}.
}
\references{

  Torsten Hothorn, Lisa Moest, Peter Buehlmann (2016), Most Likely 
  Transformations, http://arxiv.org/abs/1508.06749

}
\author{
  This package is authored by Torsten Hothorn <Torsten.Hothorn@R-project.org>.
}
\keyword{package}
